feat: update login signature handling and challenge payload in multiple services
- Refactor login method to accept challenge payload alongside signature
- Update GameSessionService and AuthService to handle new parameters
- Modify persistence methods to be asynchronous for player and match data
- Adjust ruleset handling to use Arc for better memory management
- Enhance gas management in SettlementQueue for fee estimation
- Update Solidity contracts to reflect changes in match structure and types

+++ contracts/src/AMPRegistry.sol
@@ -140,7 +140,7 @@ contract AMPRegistry is ERC2771Context, Ownable2Step, Pausable {
             playerB: address(0),
             stakeAmount: actualStake,
             state: AMPTypes.MatchState.OPEN,
-            createdAt: block.timestamp
+            createdAt: uint64(block.timestamp)
         });
 
         emit MatchCreated(matchId, gameId, _msgSender(), actualStake);
